Wrap proposal text at 160 chars
When the text wraps automatically in MarkDown it's ugly, and makes it hard to tell which words have been changed between revisions because each change appears
to change a whole paragraph. We should wrap the proposal text at 160 chars so that it lays out nicely on screen. Tested with 9pt Droid Sans Mono font on Ubuntu
with two columns open in my Vim editor at 2560 x 1440 resolution.